메인 씬에 스프라이트 애니메이션 연결 - 플레이어 텍스처와 스프라이트 애니메이션 클립 로드 추가 - MainScene에서 SpriteAnimator를 붙이고 Move 애니메이션 재생 연결 - 시작 씬을 MainScene으로 변경 - HiFi-Rush 프로젝트에 Engine 참조 추가 - 윈도우 빌드 및 줄바꿈 관련 자잘한 설정 정리

+++ Engine/Time.cpp
@@ -36,7 +36,7 @@ namespace gm
 			timeElapsed = 0.0f;
 		}
 
-		TextOut(hDC, 0, 0, szFPS, wcsnlen_s(szFPS, 50));
+		TextOut(hDC, 0, 0, szFPS, wcsnlen_s(szFPS, static_cast<size_t>(50)));
 	}
 
 	void Time::calcDeltaTime()
